+++ src/Extension/Core/Widget/BlockRenderer.php
@@ -16,9 +16,8 @@ use PhpTui\Tui\Widget\WidgetRenderer;
 
 final class BlockRenderer implements WidgetRenderer
 {
-    public function render(WidgetRenderer $renderer, Widget $widget, Buffer $buffer): void
+    public function render(WidgetRenderer $renderer, Widget $widget, Buffer $buffer, Area $area): void
     {
-        $area = $buffer->area();
         if (!$widget instanceof BlockWidget) {
             return;
         }
@@ -35,7 +34,8 @@ final class BlockRenderer implements WidgetRenderer
             $renderer->render(
                 $renderer,
                 $widget->widget,
-                $subBuffer
+                $subBuffer,
+                $innerArea
             );
             $buffer->putBuffer($innerArea->position, $subBuffer);
         }