+++ src/Extension/Core/Widget/ListRenderer.php
@@ -13,9 +13,8 @@ use PhpTui\Tui\Widget\WidgetRenderer;
 
 final class ListRenderer implements WidgetRenderer
 {
-    public function render(WidgetRenderer $renderer, Widget $widget, Buffer $buffer): void
+    public function render(WidgetRenderer $renderer, Widget $widget, Buffer $buffer, Area $area): void
     {
-        $area = $buffer->area();
         if (!$widget instanceof ListWidget) {
             return;
         }