+++ src/Widget/WidgetRenderer.php
@@ -4,9 +4,10 @@ declare(strict_types=1);
 
 namespace PhpTui\Tui\Widget;
 
+use PhpTui\Tui\Display\Area;
 use PhpTui\Tui\Display\Buffer;
 
 interface WidgetRenderer
 {
-    public function render(WidgetRenderer $renderer, Widget $widget, Buffer $buffer): void;
+    public function render(WidgetRenderer $renderer, Widget $widget, Buffer $buffer, Area $area): void;
 }