+++ tests/Unit/Model/Display/DisplayTest.php
@@ -7,12 +7,12 @@ namespace PhpTui\Tui\Tests\Unit\Model\Display;
 use PhpTui\Tui\Canvas\Marker;
 use PhpTui\Tui\Color\AnsiColor;
 use PhpTui\Tui\Display\Backend\DummyBackend;
-use PhpTui\Tui\Display\Buffer;
 use PhpTui\Tui\DisplayBuilder;
 use PhpTui\Tui\Extension\Core\Shape\PointsShape;
+use PhpTui\Tui\Extension\Core\Widget\Buffer\BufferContext;
+use PhpTui\Tui\Extension\Core\Widget\BufferWidget;
 use PhpTui\Tui\Extension\Core\Widget\CanvasWidget;
 use PhpTui\Tui\Extension\Core\Widget\ParagraphWidget;
-use PhpTui\Tui\Extension\Core\Widget\RawWidget;
 use PhpTui\Tui\Position\Position;
 use PHPUnit\Framework\TestCase;
 
@@ -25,10 +25,10 @@ final class DisplayTest extends TestCase
         $backend->setDimensions(2, 2);
 
         // intentionally go out of bounds
-        $terminal->draw(new RawWidget(function (Buffer $buffer): void {
+        $terminal->draw(new BufferWidget(function (BufferContext $context): void {
             for ($y = 0; $y < 4; $y++) {
                 for ($x = 0; $x < 4; $x++) {
-                    $buffer->putString(new Position($x, $y), 'h');
+                    $context->buffer->putString(new Position($x, $y), 'h');
                 }
             }
         }));
@@ -44,10 +44,10 @@ final class DisplayTest extends TestCase
     {
         $backend = DummyBackend::fromDimensions(4, 4);
         $terminal = DisplayBuilder::default($backend)->build();
-        $terminal->draw(new RawWidget(function (Buffer $buffer): void {
+        $terminal->draw(new BufferWidget(function (BufferContext $context): void {
             $x = 0;
             for ($y = 0; $y <= 4; $y++) {
-                $buffer->putString(new Position($x++, $y), 'x');
+                $context->buffer->putString(new Position($x++, $y), 'x');
             }
         }));
         self::assertEquals(